renamed from TableChart to Table
